my last push before going home

+++ learning/Ray_tracing.h
@@ -50,7 +50,7 @@ typedef struct color
 void  put_pixel(int x, int y, int color, t_mlx mlx);
   
 int distance_two_point(int x1, int y1, int x2, int y2);
-bool  compare_matrix(int mat1[][4], int mat2[][4], int hight, int width);
+bool  compare_matrix(float mat1[][4], float mat2[][4], int hight, int width);
 tuple new_tuple(float x, float y, float z, float w);
 void put_circle(t_mlx mlx, int color);
 void test_matrix();
